pyproject.toml
[tool.poetry]
name = "quantalpha"
version = "0.1.0"
description = "AI-powered portfolio management with real-time market analysis"
authors = ["QuantAlpha Team <team@quantalpha.com>"]
readme = "README.md"
packages = [{include = "scripts"}]

[tool.poetry.dependencies]
python = "^3.9"
groq = "^0.4.1"
python-dotenv = "^1.0.0"
numpy = "^1.24.3"
pandas = "^2.0.3"
matplotlib = "^3.7.2"
requests = "^2.31.0"
scipy = "^1.11.0"
yfinance = "^0.2.18"
plotly = "^5.17.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
flake8 = "^6.0.0"
mypy = "^1.5.0"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
portfolio-manager = "scripts.portfolio_manager:main"
ai-advisor = "scripts.ai_portfolio_advisor:main"
scenario-analysis = "scripts.scenario_analysis:main"
risk-profiler = "scripts.risk_profiler:main"
